complete_chord_identification/requirements.txt
absl-py==0.15.0
anyio==3.3.4
argon2-cffi==21.1.0
astunparse==1.6.3
async-generator==1.10
bleach==4.1.0
brotlipy==0.7.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
clang==5.0
cloudpickle==2.0.0
contextvars==2.4
cycler==0.11.0
entrypoints==0.3
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.18.2
grpcio==1.41.1
gym==0.21.0
h5py==3.1.0
idna==3.3
importlib-metadata==4.8.2
ipykernel==5.5.6
jedi==0.18.0
Jinja2==3.0.3
joblib==1.1.0
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyter-server==1.11.2
jupyterlab==3.2.3
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
mistune==0.8.4
more-itertools==8.11.0
music21==6.7.1
nbclassic==0.3.4
nbclient==0.5.5
notebook==6.4.5
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.2
pandas==1.1.5
pandocfilters==1.5.0
Pillow==8.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.22
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydotplus==2.0.2
pyrsistent==0.18.0
pytz==2021.3
pywin32==302
pywinpty==1.1.5
pyzmq==22.3.0
requests-oauthlib==1.3.0
rsa==4.7.2
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.6.0
tensorflow-gpu==2.6.2
termcolor==1.1.0
terminado==0.12.1
tqdm==4.62.3
urllib3==1.26.7
webcolors==1.11.1
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.2
wincertstore==0.2
wrapt==1.12.1
